// src/injected/selectorEngine.d.ts
/**
 * Copyright (c) Microsoft Corporation.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type SelectorType = 'default' | 'notext';
export type SelectorRoot = Element | ShadowRoot | Document;

export interface SelectorEngine {
  create(root: SelectorRoot, target: Element, type?: SelectorType): string | undefined;
  query(root: SelectorRoot, selector: string): Element | undefined;
  queryAll(root: SelectorRoot, selector: string): Element[];
}